Fotofly | Photography WordPress Theme
The main intention for Fotofly was to create a WordPress Theme for photographers, bloggers and designers that would give a maximum creative freedom for users not deeply involved in design & coding. The result is impressive with its outstanding unique elements and options. It has so many options that we even coudn’t count them. Which means it gives you full freedom to create your dream website. It’s a wordpress onestop tool for photography, and truly easy to use.